refactor: extract tui daemon launcher
